move isAllOnes() function to utils
